OSTC3: add battery percentage
From firmware version 2.15, the battery percentage is shown on the
unit (in the logbook that is, it was shown on the display for a very
long time). The used byte in the OSTC3 seems to be populated since
firmware version 2.10. The new percentage data is added to the
"battery at end" voltage data, and a V is added to denote the unit
of the voltage.

In addition, reordered the DC_FIELD_STRINGs a little, so that the
serial is on top.

Further, changed BUFLEN to 32 (as in 84eb59c) due the the length
of the new battery notation.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>