Complete the EON Steel HDLC encoding/decoding work
The previous patch did the HDLC encoding on packet send, this does the
HDLC decoding on the receive side.

In order to properly check the data integrity, the HDLC-encoded packet
needs to be fully received before it can be processed.  So while the HID
downloading continues to work packet-by-packet, the HDLC encoded BLE
GATT stream needs a temporary buffer for the data that gets filled as we
ask for the reply header.

Right now only the old USB HID path is actually tested, because I
haven't flushed out the packet receiving side in subsurface yet.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>