EON Steel: explicitly sort the dive directory listing
The code relied on the dives being ordered by date in the directory
listing, which is normally true.  But once the dive computer fills up
(after 400 dives or so), it starts re-using entries in the dive list,
and the dive directory is no longer ordered by date.

This matters because we want to download the latest dives first, since
subsurface will generally stop downloading once it finds an existing
dive.

NOTE! If you screw up the date on the dive computer, the old behavior
was possibly more to your liking, since - as long as the dive list
hadn't filled up - it wouldn't really order by date, but by dive
creation.

However, I don't see any way to get that information once the dive list
has filled up, so "order by dive date" is as good as it gets.

If you do screw up dates, and you want to download new dives that are
"older" than the dives you already have, you will need to basically set
the "download all dives" flag, and then select the new dives manually.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>