remove 'serial' number from descriptor array
I added this long ago as a merge fixup (see commit f9db4ca97c4b: "Merge
git://git.libdivecomputer.org/libdivecomputer into Subsurface-branch"),
because some of the OSTC parser_create() functions take a serial number.

However, the serial number is never actually taken from the descriptor
array, and if it was, it would be incorrect anyway.  None of the entries
in the descriptor array actually have any, so it's always zero.

So just remove it, because it's a big pain whenever upstream changes the
descriptor array format.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>