get rid of annoying compile warnings
This gets rid of the warnings in the build, mostly by just disabling
them entirely, and in one case by adding a default statement to make the
compiler not warn about lacking case statements for a switch()
statement.

Some of the warnings really don't work that well for libdivecomputer
(warning about pointer signs is very annoying when we mix regular
strings and various "unsigned char *" raw data).

And some of the warnings we should probably re-enable once by one, and
actually fix up.

But even the ones we should re-enable are right now more of a pain than
they are worth, and nobody has had the energy to do so.  And as long as
there are *so* many warnings, nobody likely will.

In the meantime, this gets rid of the harmless warnings, so that we can
see if any *bad* warnings happen.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>