remove the transfer type comments from the descriptor array
This basically reverts commit 4a3f7a7cce54 ("Mark DC descriptors with
supported transport").

The commit wasn't wrong, but it ends up having caused a *lot* of pain
for merging, because it adds the transfer type marker to most lines in
that descriptor array, and the end result is very painful to merge if
upstream libdivecomputer ever changes any of the array entries (which
happens fairly regularly, either because of format changes, or because a
new entry is added).

It's probably better and simpler to simply have some entirely separate
model for figuring out which dive computer supports which protocol.
There are enough common rules (like "old Suunto models are all FTDI")
that such a separate table probably would be simpler and clearer anyway.

This does not implement such a table, though - it just removes the
information entirely.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>